- Additional IntPtr work

+++ Source/Mosa.Kernel.x86/Thread.cs
@@ -1,5 +1,7 @@
 ﻿// Copyright (c) MOSA Project. Licensed under the New BSD License.
 
+using System;
+
 namespace Mosa.Kernel.x86
 {
 	public enum ThreadStatus { Empty = 0, Running, Terminating, Terminated, Waiting };
@@ -7,9 +9,9 @@ namespace Mosa.Kernel.x86
 	internal class Thread
 	{
 		public ThreadStatus Status = ThreadStatus.Empty;
-		public uint StackTop;
-		public uint StackBottom;
-		public uint StackStatePointer;
+		public IntPtr StackTop;
+		public IntPtr StackBottom;
+		public IntPtr StackStatePointer;
 		public uint Ticks;
 	}
 }
